+++ messages/ar/models_Tag.php
@@ -17,8 +17,7 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
-    'Description' => '',
     'ID' => '',
     'Title' => '',
-    'Url' => '',
+    'Keyword' => '',
 ];