+++ messages/de/forms.php
@@ -1,0 +1,6 @@
+<?php
+return array (
+    'Back to modules' => 'Zurück zu den Modulen',
+    'Include content tags' => 'Content-Tags integrieren',
+    '<strong>Tags</strong> module configuration' => '<strong>Tags</strong> Modul-Konfiguration',
+);