+++ messages/de/models_Tag.php
@@ -2,6 +2,5 @@
 return array (
   'ID' => 'ID',
   'Title' => 'Titel',
-  'Description' => 'Beschreibung',
-  'Url' => 'Url',
+  'Keyword' => 'Suchwort',
 );