+++ messages/de/widgets.php
@@ -1,0 +1,5 @@
+<?php
+return array (
+    'Default settings' => 'Standardeinstellungen',
+    'Custom tags' => 'Benutzerdefinierte Tags',
+);