+++ messages/ht/forms.php
@@ -1,0 +1,6 @@
+<?php
+return array (
+    'Back to modules' => '',
+    'Include content tags' => '',
+    '<strong>Tags</strong> module configuration' => '',
+);