+++ messages/id/config.php
@@ -1,0 +1,8 @@
+<?php
+return array (
+    'No custom tags created yet' => '',
+    'Here you can configure default settings for tags module' => '',
+    'Default tags settings' => '',
+    'Custom tags settings' => '',
+    'Here you can configure custom tags' => '',
+);